test: update in memory drivers repository
