--- test/respositories/in-memory-delivery-driver-repository.ts
@@ -1,53 +1,0 @@
-import { PaginationParams } from '@/core/repositories/pagination-params'
-import { DeliveryDriverRepository } from '@/domain/delivery/application/repositories/delivery-driver-repository'
-import { DeliveryDriver } from '@/domain/delivery/enterprise/entities/delivery-driver'
-
-export class InMemoryDeliveryDriverRepository extends DeliveryDriverRepository {
-  public items: DeliveryDriver[] = []
-
-  async findById(id: string): Promise<DeliveryDriver | null> {
-    const deliverDriver = this.items.find((item) => item.id.toString() === id)
-
-    if (!deliverDriver) {
-      return null
-    }
-
-    return deliverDriver
-  }
-
-  async findByTaxId(taxId: string): Promise<DeliveryDriver | null> {
-    const deliverDriver = this.items.find((item) => item.taxId === taxId)
-
-    if (!deliverDriver) {
-      return null
-    }
-
-    return deliverDriver
-  }
-
-  async findAll({ page }: PaginationParams): Promise<DeliveryDriver[]> {
-    const deliverDrivers = this.items.slice((page - 1) * 10, page * 10)
-
-    return deliverDrivers
-  }
-
-  async create(deliverDriver: DeliveryDriver) {
-    this.items.push(deliverDriver)
-  }
-
-  async save(deliverDriver: DeliveryDriver): Promise<void> {
-    const itemIndex = this.items.findIndex(
-      (item) => item.id === deliverDriver.id,
-    )
-
-    this.items[itemIndex] = deliverDriver
-  }
-
-  async delete(deliverDriver: DeliveryDriver): Promise<void> {
-    const itemIndex = this.items.findIndex(
-      (item) => item.id === deliverDriver.id,
-    )
-
-    this.items.splice(itemIndex, 1)
-  }
-}
